SUMMER: Small-Area-Estimation Unit/Area Models and Methods for Estimation in R

Provides methods for spatial and spatio-temporal smoothing of demographic and health indicators using survey data, with particular focus on estimating and projecting under-five mortality rates, described in Mercer et al. (2015) doi:10.1214/15-AOAS872 , Li et al. (2019) doi:10.1371/journal.pone.0210645 , Wu et al. (DHS Spatial Analysis Reports No. 21, 2021), and Li et al. (2023) doi:10.48550/arXiv.2007.05117 .

SUMMER provides methods for spatial and spatio-temporal smoothing of demographic and health indicators using survey data, with particular focus on estimating and projecting under-five mortality rates.

Details

For details on the model implemented in this package, see Mercer et al. (2015) doi:10.1214/15-AOAS872 and Li et al. (2019) doi:10.1371/journal.pone.0210645.

The development version of the package will be maintained on https://github.com/richardli/SUMMER.

See also

Author

Maintainer: Zehang R Li lizehang@gmail.com

Authors:

Bryan D Martin bmartin6@uw.edu

Yuan Hsiao yahsiao@uw.edu

Jessica Godwin jlg0003@uw.edu

John Paige paigejo@gmail.com

Peter Gao petergao@uw.edu

Jon Wakefield jonno@uw.edu

Samuel J Clark work@samclark.net

Geir-Arne Fuglstad geir-arne.fuglstad@ntnu.no

Andrea Riebler andrea.riebler@ntnu.no